refactor: LineDto 분리, api 명세 수정
